Feature/modal window

+++ trello/trello/views.py
@@ -1,13 +1,13 @@
-from django.shortcuts import get_object_or_404
-
 from rest_framework import status
 from rest_framework.decorators import (
     api_view,
+    authentication_classes,
     permission_classes,
 )
 
 from rest_framework.authtoken.models import Token
 from rest_framework.authtoken.views import ObtainAuthToken
+from rest_framework.authentication import SessionAuthentication, BasicAuthentication
 from rest_framework.permissions import IsAuthenticated, AllowAny
 from rest_framework.response import Response
 
@@ -175,7 +175,6 @@ def create_column(request):
 def create_card(request):
 
     card_column = request.data["column"]
-
     last_card_in_column = Card.objects.filter(column=card_column).last()
 
     if last_card_in_column:
@@ -200,20 +199,30 @@ def create_card(request):
 @api_view(["POST"])
 @permission_classes([IsAuthenticated])
 def delete_column(request):
-
-    id_column_deleted = request.data["id_column"]
-
     try:
-        Column.objects.filter(id=id_column_deleted).delete()
+        id_column_deleted = request.data["id_column"]
+        if id_column_deleted:
+            Column.objects.filter(id=id_column_deleted).delete()
     except:
         return Response(False, status=status.HTTP_404_NOT_FOUND)
+    return Response(True, status=status.HTTP_200_OK)
 
+
+@api_view(["POST"])
+@permission_classes([IsAuthenticated])
+def delete_card(request):
+    try:
+        id_card_deleted = request.data["id_card"]
+        if id_card_deleted:
+            Card.objects.filter(id=id_card_deleted).delete()
+    except:
+        return Response(False, status=status.HTTP_404_NOT_FOUND)
     return Response(True, status=status.HTTP_200_OK)
 
 
 @api_view(["GET", "POST"])
 @permission_classes([IsAuthenticated])
-def column(request):
+def take_data_column(request):
     if request.data['id']:
         column_id = request.data['id']
         queryset = Column.objects.all().filter(id=column_id)
@@ -224,19 +233,46 @@ def column(request):
 
 @api_view(["GET", "POST"])
 @permission_classes([IsAuthenticated])
-def card(request):
-    # print(request.data['id'])
-    # if request.method == "POST":
-    #     serializer = CardSerializer(data=request.data)
-    #     if serializer.is_valid():
-    #         serializer.save()
-    #         return Response(serializer.data, status=status.HTTP_201_CREATED)
-    #     return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+def take_data_card(request):
     if request.data['id']:
         card_id = request.data['id']
         queryset = Card.objects.all().filter(id=card_id)
 
     serializer = CardSerializer(queryset, many=True)
+    return Response(serializer.data)
+
+
+@api_view(["GET", "POST"])
+@permission_classes([IsAuthenticated])
+def new_data_card(request):
+    if request.data['id'] and request.data['name']:
+        card_id = request.data['id']
+        card_new_name = request.data['name']
+        try:
+            Card.objects.filter(id=card_id).update(name=card_new_name)
+        except:
+            print("если что-то сюда прилетит, то будем разбираться")
+            return Response(False, status=status.HTTP_404_NOT_FOUND)
+
+        queryset = Card.objects.all().filter(id=card_id)
+        serializer = CardSerializer(queryset, many=True)
+    return Response(serializer.data)
+
+
+@api_view(["GET", "POST"])
+@permission_classes([IsAuthenticated])
+def new_data_column(request):
+    if request.data['id'] and request.data['name']:
+        column_id = request.data['id']
+        column_new_name = request.data['name']
+        try:
+            Column.objects.filter(id=column_id).update(name=column_new_name)
+        except:
+            print("если что-то сюда прилетит, то будем разбираться")
+            return Response(False, status=status.HTTP_404_NOT_FOUND)
+
+        queryset = Column.objects.all().filter(id=column_id)
+        serializer = ColumnSerializer(queryset, many=True)
     return Response(serializer.data)
